Updated routes for custodies and header component styles
